requirements.txt
apeye>=0.3.0
click>=7.1.2
consolekit>=0.3.1
domdf-python-tools>=1.5.0
dulwich!=0.20.7,!=0.20.8,>=0.20.5
pygithub!=1.54,>=1.51
shippinglabel>=0.6.0
typing-extensions>=3.7.4.3
